+++ src/json_write_from_acmap.cpp
@@ -358,6 +358,7 @@ std::string acmap_to_json(
   // = OTHER =
   if (!map.isdefault("ag_group_levels"))   x.AddMember("agv", jsonifya(map.get_ag_group_levels(), allocator), allocator);
   if (!map.isdefault("sr_group_levels"))   x.AddMember("srv", jsonifya(map.get_sr_group_levels(), allocator), allocator);
+  if (!map.isdefault("layer_names"))       x.AddMember("ln",  jsonifya(map.get_layer_names(), allocator), allocator);
   if (!map.isdefault("dilution_stepsize")) x.AddMember("ds",  map.dilution_stepsize, allocator);
   if (!map.isdefault("description"))       x.AddMember("D",   jsonifya(map.description, allocator), allocator);