+++ tests/testthat/test-map_merging.R
@@ -161,7 +161,7 @@ test_that("Merging titers", {
   map13 <- mergeMaps(list(map1, map3))
   # expect_equal(unname(titerTable(map13)), matrix(logtoraw(-1:4+2), 3, 2))
   expect_equal(unname(titerTable(map13)), matrix("*", 3, 2))
-  expect_equal(titerTableLayers(map13), list(
+  expect_equal(unname(titerTableLayers(map13)), list(
     matrix(logtoraw(-1:4 + 1), 3, 2),
     matrix(logtoraw(-1:4 + 3), 3, 2)
   ))
@@ -385,6 +385,103 @@ test_that("Merging serum and antigen groups", {
   expect_equal(
     as.character(srGroups(merged_map)),
     sr_groups[match(srNames(merged_map), sr_names)]
+  )
+
+})
+
+# Merging maps with names
+test_that("Merging maps with names", {
+
+  mergemap1a <- mergemap1
+  mergemap2a <- mergemap2
+
+  mapName(mergemap1a) <- "Merge map 1"
+  mapName(mergemap2a) <- "Merge map 2"
+
+  merged_map_unnamed <- mergeMaps(list(mergemap1, mergemap2))
+  merged_map_named <- mergeMaps(list(mergemap1a, mergemap2a))
+
+  # Check null defaults
+  expect_null(layerNames(mergemap1))
+  expect_null(layerNames(mergemap1a))
+
+  expect_equal(
+    layerNames(merged_map_unnamed),
+    c("", "")
+  )
+  expect_equal(
+    names(titerTableLayers(merged_map_unnamed)),
+    c("", "")
+  )
+
+  # Check merge results
+  expect_equal(
+    names(titerTableLayers(merged_map_named)),
+    c("Merge map 1", "Merge map 2")
+  )
+  expect_equal(
+    layerNames(merged_map_named),
+    c("Merge map 1", "Merge map 2")
+  )
+
+  # Check changing names
+  layerNames(merged_map_unnamed) <- c("Merge map 1a", "Merge map 2a")
+
+  expect_equal(
+    names(titerTableLayers(merged_map_unnamed)),
+    c("Merge map 1a", "Merge map 2a")
+  )
+  expect_equal(
+    layerNames(merged_map_unnamed),
+    c("Merge map 1a", "Merge map 2a")
+  )
+
+  # Check removing names
+  layerNames(merged_map_unnamed) <- NULL
+
+  expect_equal(
+    layerNames(merged_map_unnamed),
+    c("", "")
+  )
+  expect_equal(
+    names(titerTableLayers(merged_map_unnamed)),
+    c("", "")
+  )
+
+  # Check errors
+  expect_error({
+    layerNames(merged_map_unnamed) <- "Merge map 1a"
+  })
+
+  # Check saving and loading
+  tmp <- tempfile(fileext = ".ace")
+  save.acmap(merged_map_named, tmp)
+  merged_map_loaded <- read.acmap(tmp)
+
+  expect_equal(
+    names(titerTableLayers(merged_map_loaded)),
+    c("Merge map 1", "Merge map 2")
+  )
+  expect_equal(
+    layerNames(merged_map_loaded),
+    c("Merge map 1", "Merge map 2")
+  )
+
+  # Take layer names from list names when merging
+  merged_map_listnamed <- mergeMaps(
+    list(
+      map1merge = mergemap1a,
+      map2merge = mergemap2a
+    )
+  )
+
+  expect_equal(
+    names(titerTableLayers(merged_map_listnamed)),
+    c("map1merge", "map2merge")
+  )
+  expect_equal(
+    layerNames(merged_map_listnamed),
+    c("map1merge", "map2merge")
   )
 
 })